+++ ReleaseNotes.md
@@ -10,6 +10,7 @@
 
 ## Bug fixes
 - cta/CTA#1092 - Fix overflow error with drive state latestBandwith causing cta frontend crash
+- cta/CTA#501 - Fix disappearing reason when TapeDrive is reading or writing
 
 # v4.4.0-1