+++ ReleaseNotes.md
@@ -13,6 +13,15 @@
 - cta/CTA#1092 - Fix overflow error with drive state latestBandwith causing cta frontend crash
 - cta/CTA#501 - Fix disappearing reason when TapeDrive is reading or writing
 
+# v4.4.1-1
+
+## Upgrade Instructions
+
+## Features
+
+## Bug fixes
+- cta/CTA#1092 Fix overflow with drive bandwith causing frontend to crash
+
 # v4.4.0-1
 
 ## Upgrade Instructions